move diff hist
